+++ photons_app/__init__.py
@@ -1,4 +1,4 @@
-VERSION = "0.5.6"
+VERSION = "0.5.7"
 
 __shortdesc__ = """Base module for all photons applications"""